+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Changed
 - 🧨 Lyo is now built using JDK 11
-- 🧨 Jena is upgraded to 4.4.0
+- 🧨 Jena is upgraded to 4.5.0 **(addresses [CVE-2022-28890](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-28890))**
   - Jena renamed `RDFReader/RDFWriter` to `RDFReaderI/RDFWriterI`
 - LyoStore: Ordering resources by their subject IDs when doing a query to store. This ordering can be disabled with a call to `OSLC4JUtils.setLyoStorePagingUnsafe(true)`
 - LyoStore: `OSLC4JUtils.hasLyoStorePagingPreciseLimit()` will return true by default. Call `OSLC4JUtils.setLyoStorePagingPreciseLimit(false)` to restore the old behavior.