+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - 🧨 Lyo is now built using JDK 11
 - 🧨 Jena is upgraded to 4.5.0 **(addresses [CVE-2022-28890](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-28890))**
   - Jena renamed `RDFReader/RDFWriter` to `RDFReaderI/RDFWriterI`
+- 🧨 TRS now uses BigInteger instead of 32-bit ints for `trs:order` properties, in line with the spec. 
 - LyoStore: Ordering resources by their subject IDs when doing a query to store. This ordering can be disabled with a call to `OSLC4JUtils.setLyoStorePagingUnsafe(true)`
 - LyoStore: `OSLC4JUtils.hasLyoStorePagingPreciseLimit()` will return true by default. Call `OSLC4JUtils.setLyoStorePagingPreciseLimit(false)` to restore the old behavior.
 - `oslc4j-json4j-provider` uses `wink-json4j` version 1.4 instead of 1.2.1-incubating.