+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Eclipse Lyo changelog
 
-## [Unreleased]
+## [5.0.0]
 
 ### Added